packages/gbp-infra/pyproject.toml
[project]
name = "gbp-infra"
version = "0.1.0"
description = "AWS CDK infrastructure for the Geospatial Benchmarking Protocol"
readme = "README.md"
authors = [
    { name = "Pete Gadomski", email = "pete.gadomski@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.200",
    "constructs>=10",
    "pydantic-settings>=2",
]

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
